+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Return a successful exit code when Misti analysis does not generate any warnings
+
 ## [0.4.1] - 2024-10-12
 
 ### Changed